refactor(WebAPI): fixed an issue where some parts of the files were skipped

+++ tests/BaseTests/FakeData/FakeData.FileData.cs
@@ -4,12 +4,13 @@ namespace PlexRipper.BaseTests;
 
 public static partial class FakeData
 {
-    public static byte[] GetDownloadFile(int sizeInMib)
+    public static byte[] GetDownloadFile(double sizeInMib)
     {
-        var b = new byte[(long)ByteSize.FromMebiBytes(sizeInMib).Bytes]; // convert mib to byte
+        // convert mib to byte
+        var b = new byte[(long)ByteSize.FromMebiBytes(sizeInMib).Bytes];
         RandomInstance.NextBytes(b);
         return b;
     }
 
-    public static MemoryStream GetFileStream(int sizeInMib) => new(GetDownloadFile(sizeInMib));
+    public static MemoryStream GetFileStream(double sizeInMib) => new(GetDownloadFile(sizeInMib));
 }
